Changes to reflect the integration with x10.lang API.
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@224 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src/x10/array/point_c.java
@@ -20,14 +20,14 @@ public class point_c extends point {
 	
 	/** Return the value of this point on the i'th dimension.
 	 */    
-	public int valueAt( /*nat*/long i ) {
+	public int valueAt( /*nat*/int i ) {
 		return val[((int) i) % val.length];
 	}
 	
 	/** Return true iff the point is on the upper boundary of the i'th
 	 * dimension of its region.
 	 */
-	public boolean onUpperBoundary( /*nat*/long i ) {
+	public boolean onUpperBoundary( /*nat*/int i ) {
 		return region.rank(i).high() == valueAt(i);
 	}
 	
@@ -35,7 +35,7 @@ public class point_c extends point {
 	/** Return true iff the point is on the lower boundary of the i'th
 	 * dimension.
 	 */
-	public boolean onLowerBoundary( /*nat*/long i ) {
+	public boolean onLowerBoundary( /*nat*/int i ) {
 		return region.rank(i).low() == valueAt(i);
 	}
